+++ storage/storage.go
@@ -6,8 +6,8 @@ import (
 type Storage interface {
 	Save(p *config.Configuration) error
 	//Update(p *config.Configuration) error
-	GetConfig(chatId int) (*config.Configuration, error)
-	IsExists(chatId int) (bool, error)
+	GetConfig(chatId int64) (*config.Configuration, error)
+	IsExists(chatId int64) (bool, error)
 }
